Switch u8 pixels to i16 coefficients
Fix quantization so that it works with i16 coefficients.

Temporarily disable transform simd.

+++ src/util.rs
@@ -198,7 +198,7 @@ pub trait Pixel:
 }
 
 impl Pixel for u8 {
-  type Coeff = i32;
+  type Coeff = i16;
 
   fn type_enum() -> PixelType {
     PixelType::U8
